fix!: Remove `git-config` test utilities from `git-path`.
