+++ git-path/Cargo.toml
@@ -15,5 +15,4 @@ doctest = false
 
 [dependencies]
 bstr = { version = "0.2.17", default-features = false, features = ["std"] }
-tempfile = "3.2.0"
 thiserror = "1.0.26"
